python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
redis==5.0.1
structlog==23.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
pillow==10.1.0
